requirements.txt
python-dotenv 1.0.1
requests 2.32.3
fitz 0.0.1.dev2
frontend 0.0.3
pymupdf 1.24.10